IPL-AI-Predictor/requirements.txt
pandas
scikit-learn
fastapi
uvicorn
streamlit
requests
python-dotenv